src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-root-innerHTML-expected.txt

CONSOLE ERROR: line 17: Uncaught TypeError: host.createShadowRoot is not a function
This tests the innerHTML property of a shadow root.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
